requirements.txt
PyYAML==5.4.1
requests==2.26.0
RPi.GPIO==0.7.0
spidev==3.5
